+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "budget",
+  "name": "budgeting_group_app",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -9,14 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-community/datetimepicker": "^7.6.2",
     "expo": "50.0.2",
     "expo-status-bar": "1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.9",
+    "@babel/core": "7.20.0",
     "@types/react": "18.2.45",
     "typescript": "5.1.3"
   },
